Re-namespace users client tests
