+++ tests/Functional/Client/UsersClient/AbstractUsersClientTestCase.php
@@ -1,0 +1,38 @@
+<?php
+
+declare(strict_types=1);
+
+namespace SmartAssert\ApiClient\Tests\Functional\Client\UsersClient;
+
+use GuzzleHttp\Psr7\HttpFactory;
+use SmartAssert\ApiClient\Tests\Functional\Client\AbstractClientTestCase;
+use SmartAssert\ApiClient\Tests\Functional\DataProvider\CommonNonSuccessResponseDataProviderTrait;
+use SmartAssert\ApiClient\UsersClient;
+use SmartAssert\ServiceClient\Client as ServiceClient;
+use SmartAssert\ServiceClient\ExceptionFactory\CurlExceptionFactory;
+use SmartAssert\ServiceClient\ResponseFactory\ResponseFactory;
+
+abstract class AbstractUsersClientTestCase extends AbstractClientTestCase
+{
+    use CommonNonSuccessResponseDataProviderTrait;
+
+    protected UsersClient $client;
+
+    protected function setUp(): void
+    {
+        parent::setUp();
+
+        $httpFactory = new HttpFactory();
+
+        $this->client = new UsersClient(
+            'https://users.example.com',
+            new ServiceClient(
+                $httpFactory,
+                $httpFactory,
+                $this->httpClient,
+                ResponseFactory::createFactory(),
+                new CurlExceptionFactory(),
+            ),
+        );
+    }
+}
