+++ tests/Functional/Client/UsersClient/GetUserApiKeyTest.php
@@ -2,20 +2,20 @@
 
 declare(strict_types=1);
 
-namespace SmartAssert\ApiClient\Tests\Functional\Client;
+namespace SmartAssert\ApiClient\Tests\Functional\Client\UsersClient;
 
 use GuzzleHttp\Psr7\Response;
-use SmartAssert\ApiClient\Model\User;
+use SmartAssert\ApiClient\Model\ApiKey;
 use SmartAssert\ApiClient\Tests\Functional\DataProvider\InvalidJsonResponseExceptionDataProviderTrait;
 use SmartAssert\ApiClient\Tests\Functional\DataProvider\NetworkErrorExceptionDataProviderTrait;
 use SmartAssert\ServiceClient\Exception\InvalidModelDataException;
 
-class CreateUserTest extends AbstractClientTestCase
+class GetUserApiKeyTest extends AbstractUsersClientTestCase
 {
     use InvalidJsonResponseExceptionDataProviderTrait;
     use NetworkErrorExceptionDataProviderTrait;
 
-    public function testCreateUserThrowsInvalidModelDataException(): void
+    public function testGetUserApiKeyThrowsInvalidModelDataException(): void
     {
         $responsePayload = ['key' => 'value'];
         $response = new Response(200, ['content-type' => 'application/json'], (string) json_encode($responsePayload));
@@ -23,39 +23,38 @@ class CreateUserTest extends AbstractClientTestCase
         $this->mockHandler->append($response);
 
         try {
-            $this->client->createUser('admin token', 'user identifier', 'password');
+            $this->client->getUserApiKey('token');
             self::fail(InvalidModelDataException::class . ' not thrown');
         } catch (InvalidModelDataException $e) {
-            self::assertSame(User::class, $e->class);
+            self::assertSame(ApiKey::class, $e->class);
             self::assertSame($response, $e->response);
             self::assertSame($responsePayload, $e->payload);
         }
     }
 
-    public function testCreateUserRequestProperties(): void
+    public function testGetUserApiKeyRequestProperties(): void
     {
-        $id = md5((string) rand());
-        $userIdentifier = md5((string) rand());
+        $label = null;
+        $key = md5((string) rand());
 
         $this->mockHandler->append(new Response(
             200,
             ['content-type' => 'application/json'],
             (string) json_encode([
-                'user' => [
-                    'id' => $id,
-                    'user-identifier' => $userIdentifier,
+                'api_key' => [
+                    'label' => $label,
+                    'key' => $key,
                 ],
             ])
         ));
 
-        $adminToken = md5((string) rand());
-        $password = md5((string) rand());
+        $token = md5((string) rand());
 
-        $this->client->createUser($adminToken, $userIdentifier, $password);
+        $this->client->getUserApiKey($token);
 
         $request = $this->getLastRequest();
-        self::assertSame('POST', $request->getMethod());
-        self::assertSame('Bearer ' . $adminToken, $request->getHeaderLine('authorization'));
+        self::assertSame('GET', $request->getMethod());
+        self::assertSame('Bearer ' . $token, $request->getHeaderLine('authorization'));
     }
 
     public static function clientActionThrowsExceptionDataProvider(): array
@@ -69,7 +68,7 @@ class CreateUserTest extends AbstractClientTestCase
     protected function createClientActionCallable(): callable
     {
         return function () {
-            $this->client->createUser('admin token', 'user identifier', 'password');
+            $this->client->getUserApiKey('token');
         };
     }
 }
